requirements.txt
APScheduler==3.11.0
beautifultable==1.1.0
blinker==1.9.0
boto3==1.35.99
botocore==1.35.99
click==8.1.8
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
pysondb==1.6.7
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.17.0
tzlocal==5.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
